Adds helpers and interface modules from lcdb-workflows
I am thinking it makes sense to migrate the modules over to this
package. I also moved tests over and have them mostly working. The
helpers module has a couple of tests that require the 'wrappers' folder
from lcdb-workflow. I am thinking about changing these to use the
github.com repository for this.

The doctests in interface.py are not currently working.